requirements.txt
Django==3.1
djangorestframework==3.12.3
psycopg2-binary==2.8.6
django-rest-auth==0.9.5
django-allauth==0.44.0
django-rest-swagger==2.2.0
django-filter==2.4.0
Pillow==8.2.0
drf-yasg==1.20.0
